Ensure flow direciton raster is a float type by casting to one.
Former-commit-id: 06a27a2cc017ed11325aac90b2d3d8907520babc

+++ invest_cython_core/invest_cython_core.pyx
@@ -762,7 +762,8 @@ def flow_accumulation_dinf(flowDirection, flowAccumulation):
     #on gdal arrays, so we invert the x and y offsets here
     cdef np.ndarray[np.float_t,ndim=2] flowDirectionMatrix = \
         flowDirection.GetRasterBand(1).ReadAsArray(0, 0,
-        flowDirection.RasterXSize, flowDirection.RasterYSize).transpose()
+        flowDirection.RasterXSize, flowDirection.RasterYSize).transpose().astype(np.float)
+        
     nodataFlowDirection = flowDirection.GetRasterBand(1).GetNoDataValue()
     nodataFlowAccumulation = flowAccumulation.GetRasterBand(1).GetNoDataValue()
     gp = flowDirection.GetGeoTransform()
